refactor(modules): extract data-lifecycle bounded context
Moves cleanup retention concern into src/modules/data-lifecycle/.

- src/entities/cleanup/                                    → modules/data-lifecycle/entities/cleanup/
- src/usecases/cleanup/                                    → modules/data-lifecycle/usecases/cleanup/
- src/interface-adapters/controllers/http/cleanup.routes.ts → modules/data-lifecycle/interface-adapters/controllers/http/
- src/interface-adapters/gateways/reviewLogFile.gateway.ts  → modules/data-lifecycle/interface-adapters/gateways/
- src/interface-adapters/gateways/fileSystem/reviewLogFile.fileSystem.gateway.ts → modules/data-lifecycle/interface-adapters/gateways/fileSystem/

8 importers updated. 2 relative imports in src/main/dependencies.ts
normalized to @/ alias (project rule).
