+++ src/frameworks/scheduler/cleanupScheduler.ts
@@ -1,7 +1,7 @@
 import type { Logger } from 'pino';
 import type { ReviewFileGateway } from '@/interface-adapters/gateways/reviewFile.gateway.js';
-import type { ReviewLogFileGateway } from '@/interface-adapters/gateways/reviewLogFile.gateway.js';
-import { cleanupExpiredReviews } from '@/usecases/cleanup/cleanupExpiredReviews.usecase.js';
+import type { ReviewLogFileGateway } from '@/modules/data-lifecycle/interface-adapters/gateways/reviewLogFile.gateway.js';
+import { cleanupExpiredReviews } from '@/modules/data-lifecycle/usecases/cleanup/cleanupExpiredReviews.usecase.js';
 import { getProjectRetentionDays } from '@/config/projectConfig.js';
 
 const TWENTY_FOUR_HOURS_IN_MILLISECONDS = 86_400_000;
