+++ src/main/dependencies.ts
@@ -2,14 +2,14 @@ import type { Config } from '../config/loader.js';
 import type { ReviewRequestTrackingGateway } from '../interface-adapters/gateways/reviewRequestTracking.gateway.js';
 import type { StatsGateway } from '../interface-adapters/gateways/stats.gateway.js';
 import type { ReviewFileGateway } from '../interface-adapters/gateways/reviewFile.gateway.js';
-import type { ReviewLogFileGateway } from '../interface-adapters/gateways/reviewLogFile.gateway.js';
+import type { ReviewLogFileGateway } from '@/modules/data-lifecycle/interface-adapters/gateways/reviewLogFile.gateway.js';
 import type { ReviewContextGateway } from '../entities/reviewContext/reviewContext.gateway.js';
 import type { InsightsGateway } from '@/entities/insight/insights.gateway.js';
 import { FileSystemReviewRequestTrackingGateway } from '../interface-adapters/gateways/fileSystem/reviewRequestTracking.fileSystem.js';
 import { FileSystemStatsGateway } from '../interface-adapters/gateways/fileSystem/stats.fileSystem.js';
 import { FileSystemInsightsGateway } from '@/interface-adapters/gateways/fileSystem/insights.fileSystem.js';
 import { FileSystemReviewFileGateway } from '../interface-adapters/gateways/fileSystem/reviewFile.fileSystem.js';
-import { FileSystemReviewLogFileGateway } from '../interface-adapters/gateways/fileSystem/reviewLogFile.fileSystem.gateway.js';
+import { FileSystemReviewLogFileGateway } from '@/modules/data-lifecycle/interface-adapters/gateways/fileSystem/reviewLogFile.fileSystem.gateway.js';
 import { ReviewContextFileSystemGateway } from '../interface-adapters/gateways/reviewContext.fileSystem.gateway.js';
 import { ReviewContextWatcherService } from '../services/reviewContextWatcher.service.js';
 import { ReviewContextProgressPresenter } from '../interface-adapters/presenters/reviewContextProgress.presenter.js';
