+++ src/modules/data-lifecycle/interface-adapters/gateways/reviewLogFile.gateway.ts
