+++ src/modules/data-lifecycle/usecases/cleanup/cleanupExpiredReviews.usecase.ts
@@ -1,6 +1,6 @@
 import type { ReviewFileGateway } from '@/interface-adapters/gateways/reviewFile.gateway.js';
-import type { ReviewLogFileGateway } from '@/interface-adapters/gateways/reviewLogFile.gateway.js';
-import { RetentionPolicy } from '@/entities/cleanup/retentionPolicy.valueObject.js';
+import type { ReviewLogFileGateway } from '@/modules/data-lifecycle/interface-adapters/gateways/reviewLogFile.gateway.js';
+import { RetentionPolicy } from '@/modules/data-lifecycle/entities/cleanup/retentionPolicy.valueObject.js';
 
 export interface CleanupResult {
   deletedReviewFiles: string[];
