+++ src/tests/stubs/reviewLogFile.stub.ts
@@ -1,5 +1,5 @@
 import { join } from 'node:path';
-import type { ReviewLogFileGateway, LogFileInfo } from '@/interface-adapters/gateways/reviewLogFile.gateway.js';
+import type { ReviewLogFileGateway, LogFileInfo } from '@/modules/data-lifecycle/interface-adapters/gateways/reviewLogFile.gateway.js';
 
 interface LogFileEntry {
   mtime: string;
